change some forgoten variables from float to double. bunch of bug testing. changed the new ray direction rotation method. We now rotate one of the vectors of the triangle directly instead of finding an orthogonal vector. lowered brdfs drastically

+++ OJGIR/Ray.h
@@ -12,11 +12,11 @@ public:
 
 	Ray();
 
-	Ray(glm::dvec3 _origin, glm::dvec3 _direction, Ray* _parent, std::vector<Mesh*>* _sceneData, RNG& _rng, glm::vec3 _W);
+	Ray(glm::dvec3 _origin, glm::dvec3 _direction, Ray* _parent, std::vector<Mesh*>* _sceneData, RNG& _rng, glm::dvec3 _W);
 
 	~Ray();
 
-	void Intersection(glm::vec3 _origin, glm::vec3 _direction);
+	void Intersection(glm::dvec3 _origin, glm::dvec3 _direction);
 	void Reflection( float _u1, float _u2);
 	void Transmision();
